swell-im: move swell-im.h definitions into swell-internal.h
