+++ WDL/swell/swell-internal.h
@@ -1391,4 +1391,31 @@ static WDL_STATICFUNC_UNUSED int ext_valid_for_extlist(const char *thisext, cons
   return -1;
 }
 
+
+#ifdef SWELL_SUPPORT_IM
+
+//-----------------------------------------------------------------------------
+// Input Method Core Functions
+//-----------------------------------------------------------------------------
+void im_init(HWND hwnd);
+void im_free(HWND hwnd);
+bool im_hwnd_match(HWND hwnd, const char* classname);
+void im_update_candidates_location(HWND hwnd);
+
+
+//-----------------------------------------------------------------------------
+// Pre-edit Text Handling
+//-----------------------------------------------------------------------------
+struct ImPreeditPaintResult {
+    int sel1;
+    int sel2;
+    int cursor_pos;
+    WDL_FastString title;
+};
+
+ImPreeditPaintResult im_preedit_paint(struct __SWELL_editControlState *es, HWND hwnd,
+                                     int sel1, int sel2, int cursor_pos);
+
+#endif // SWELL_SUPPORT_IM
+
 #endif
